# apps/python/app/utils/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/nats/test": {
      "post": {
        "description": "Send a test message over NATS to the Python service",
        "operationId": "NatsController_testNatsConnection",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Test message sent successfully"
          },
          "500": {
            "description": "Failed to send test message"
          }
        },
        "summary": "Test NATS connection",
        "tags": [
          "nats"
        ]
      }
    },
...
    "/api/profile": {
      "get": {
        "description": "Get the current user's extended profile information",
        "operationId": "ProfileController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "1",
                      "userId": "1",
                      "bio": "Software developer passionate about web technologies",
                      "avatarUrl": "https://example.com/avatar.jpg",
                      "socialLinks": {
                        "twitter": "https://twitter.com/username",
                        "github": "https://github.com/username",
                        "linkedin": "https://linkedin.com/in/username",
                        "website": "https://example.com"
                      },
                      "createdAt": "2023-01-01T00:00:00.000Z",
                      "updatedAt": "2023-01-01T00:00:00.000Z"
                    },
                    "status": 200,
                    "message": "Profile retrieved successfully"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": null,
                    "status": 401,
                    "message": "Authentication required",
                    "error": {
                      "code": "UNAUTHORIZED",
                      "message": "Authentication required"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get user profile",
        "tags": [
          "profile"
        ]
      },
      "patch": {
        "description": "Update the current user's extended profile information",
        "operationId": "ProfileController_updateProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "1",
                      "userId": "1",
                      "bio": "Updated bio",
                      "avatarUrl": "https://example.com/new-avatar.jpg",
                      "socialLinks": {
                        "twitter": "https://twitter.com/newusername",
                        "github": "https://github.com/newusername"
                      },
                      "createdAt": "2023-01-01T00:00:00.000Z",
                      "updatedAt": "2023-01-02T00:00:00.000Z"
                    },
                    "status": 200,
                    "message": "Profile updated successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update user profile",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/avatar": {
      "post": {
        "description": "Upload a new avatar image for the current user",
        "operationId": "ProfileController_uploadAvatar",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Upload avatar",
        "tags": [
          "profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fullstack Boilerplate API",
    "description": "The API documentation for the Fullstack Boilerplate project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "Cookie",
        "name": "connect.sid"
      }
    },
...
  }
}